feat: Reservation Service implemented
Reservation service is on. To use the builder, modified Reservation entity and made constructor and used builder annotation there.
Has basic CRUD (except update) for user reservation. Update not implemented, because who ables people to update there reservation anyway.
For Admin method, we only have getReservationByEventId, to see all reservation.Reason we did this is because.. well it would make catastrophe problem when admin accidently erase or update a reservation. And there does not seem to be any good reason why they should be able to anyway.