Refactor: getReservation now take only phone number for it's parameter, rather than id
so I just realised that it's impposible to send schedule or event id when tryting to get reservation. so now we simply take phone number of the user, and only returns events that are 'open'. any closed events will be ignored. The result is exactly the same, but this time, it works.