Refactor: Upload system refactored to now track stored files via entity
We want to track, and find files within upload domain, so it is now renamed to storage, and has entity, repository. no new feature at this point, considering we are making session domain for controle tower that is going to track which semester it is at this time. This is just a simple refactor to bascially use new response dto, and save specifics of files in db.